janasunwai_360/requirements.txt
Flask>=3.0
python-dotenv>=1.0
supabase>=1.0
werkzeug>=3.0# Core framework
fastapi
uvicorn

# Templating & static files
jinja2
aiofiles

# Database
sqlalchemy

# Validation & settings
pydantic
python-dotenv

# Security & auth
passlib[bcrypt]
python-jose[cryptography]

# HTTP & utilities
requests

# Data & AI
pandas

# AI API
groq
# Optional: For development and testing
pytest
# Core framework
fastapi
uvicorn

# Templating & static files
jinja2
aiofiles

# Database
sqlalchemy

# Validation & settings
pydantic
python-dotenv

# Security & auth
passlib[bcrypt]
python-jose[cryptography]

# HTTP & utilities
requests

# Data & AI
pandas

# AI API
groq

# Optional: For development and testing
pytest